requirements.txt
# Core Dependencies
sentence-transformers>=3.0.0
faiss-cpu==1.13.2
rank-bm25==0.2.2
transformers>=4.40.0
torch==2.8.0
hf_xet>=0.1.0

# Data Collection
wikipedia-api>=0.7.0
beautifulsoup4>=4.12.0
requests>=2.31.0

# NLP and Text Processing
nltk>=3.9.0
spacy>=3.8.0
tiktoken>=0.8.0
# Run: python -m spacy download en_core_web_sm

# Evaluation Metrics
rouge-score>=0.1.2
bert-score>=0.3.13
scikit-learn>=1.5.0
scipy>=1.13.0

# Vector Database
chromadb>=0.5.0

# LLM and Generation
openai>=1.50.0
langchain>=0.3.0
langchain-community>=0.3.0

# UI Framework
streamlit>=1.40.0
gradio>=4.44.0

# Visualization and Reporting
matplotlib>=3.9.0
seaborn>=0.13.0
plotly>=5.24.0
pandas>=2.2.0
numpy>=2.0.0

# Report Generation
fpdf2>=2.8.0
jinja2>=3.1.0
markdown>=3.7.0

# Utilities
tqdm>=4.66.0
python-dotenv>=1.0.0
pyyaml>=6.0.0

xf_xet